src/main/java/web/lose/LoseFirstServlet.java
package web.lose;

import models.AnswerChallenge;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import service.FirstService;

import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

@WebServlet("/losefirst")
public class LoseFirstServlet extends HttpServlet {
    private static final Logger LOGGER = LoggerFactory.getLogger(LoseFirstServlet.class);
    FirstService firstService = new FirstService();
    @Override
    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        LOGGER.info("Method doPost() LoseFirstServlet started");
        AnswerChallenge answer = firstService.call(Boolean.parseBoolean(req.getParameter("answer")));

        resp.setStatus(201);
        req.setAttribute("answer", answer.getMessage());
        req.getRequestDispatcher(answer.getPage()).forward(req, resp);
        LOGGER.info("Method doPost() LoseFirstServlet finished");
    }
}
